Refactored tests. Removed duplicate unordered list test.
