+++ src/test/java/markdown_renderer/MarkdownRendererAbstractTest.java
@@ -1,0 +1,28 @@
+package markdown_renderer;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
+
+import java.io.IOException;
+
+import org.junit.Before;
+
+public abstract class MarkdownRendererAbstractTest {
+
+	private static final String RESOURCE_LOCATION = "src/test/resources/commonmark/";
+	private MarkdownRenderer markdownRenderer;
+
+	@Before
+	public void setUp() {
+		markdownRenderer = new MarkdownRenderer();
+	}
+
+	protected void readAndAssertFileContents(String filename, String contents) {
+		try {
+			String buffer = FileReaderHelper.readFileToBuffer(RESOURCE_LOCATION + filename);
+			assertEquals(buffer, markdownRenderer.render(contents));
+		} catch (IOException e) {
+			fail();
+		}
+	}
+}
